requirements.txt
google-generativeai
cohere
python-dotenv
requests